# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NGO Donor Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sponsorships:
    get:
      operationId: listSponsorships
      tags: [sponsorships]
      summary: List all sponsorships
      responses:
        "200":
          description: List of sponsorships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsorship"
    post:
      operationId: createSponsorship
      tags: [sponsorships]
      summary: Create a sponsorship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSponsorshipBody"
      responses:
        "201":
          description: Created sponsorship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsorship"
  /sponsorships/{id}:
    get:
      operationId: getSponsorship
      tags: [sponsorships]
      summary: Get sponsorship by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sponsorship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsorship"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSponsorship
      tags: [sponsorships]
      summary: Update a sponsorship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSponsorshipBody"
      responses:
        "200":
          description: Updated sponsorship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsorship"
    delete:
      operationId: deleteSponsorship
      tags: [sponsorships]
      summary: Delete a sponsorship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── REPORTS ─────────────────────────────────────────────────────────────────
...
  /timemachine:
    get:
      operationId: listTimeMachineEntries
      tags: [timemachine]
      summary: List all Time Machine entries
      responses:
        "200":
          description: List of Time Machine entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeMachineEntry"
    post:
      operationId: createTimeMachineEntry
      tags: [timemachine]
      summary: Create a Time Machine entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeMachineBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeMachineEntry"
  /timemachine/{id}:
    get:
      operationId: getTimeMachineEntry
      tags: [timemachine]
      summary: Get a Time Machine entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Time Machine entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeMachineEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTimeMachineEntry
      tags: [timemachine]
      summary: Update a Time Machine entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeMachineBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeMachineEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTimeMachineEntry
      tags: [timemachine]
      summary: Delete a Time Machine entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /timemachine/{id}/photos:
    post:
      operationId: uploadTimeMachinePhotos
      tags: [timemachine]
      summary: Upload photos for a Time Machine entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Photos uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadPhotosResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UploadPhotosResponse:
      type: object
      properties:
        photoUrls:
          type: array
          items:
            type: string
      required:
        - photoUrls

    # ─── DONOR ──────────────────────────────────────────────────────────────────
...
    Sponsorship:
      type: object
      properties:
        id:
          type: string
        donorId:
          type: string
        beneficiaryId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        monthlyAmount:
          type: number
        currency:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - donorId
        - beneficiaryId
        - startDate
        - monthlyAmount
        - currency
        - status
        - createdAt
        - updatedAt
    CreateSponsorshipBody:
      type: object
      properties:
        donorId:
          type: string
        beneficiaryId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        monthlyAmount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [Active, Paused, Ended]
        notes:
          type: ["string", "null"]
      required:
        - donorId
        - beneficiaryId
        - startDate
        - monthlyAmount
        - currency
        - status
    UpdateSponsorshipBody:
      type: object
      properties:
        endDate:
          type: ["string", "null"]
          format: date-time
        monthlyAmount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [Active, Paused, Ended]
        notes:
          type: ["string", "null"]

    # ─── REPORTS ─────────────────────────────────────────────────────────────────
...
    TimeMachineEntry:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        home:
          type: string
        eventDate:
          type: string
          format: date-time
        photos:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - home
        - eventDate
        - photos
        - createdAt
        - updatedAt
    CreateTimeMachineBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
            - Success Story
            - Inspiring Story
            - Recognition
            - Donor Support
            - Event by Kids
            - Visitor Visit
            - Challenge / Problem
            - General Update
        home:
          type: string
          enum:
            - All Homes
            - Girls Home - Uppal
            - Blind Home - Begumpet
            - Old Age Home - Peerzadiguda
        eventDate:
          type: string
          format: date-time
      required:
        - title
        - description
        - category
        - home
        - eventDate
    UpdateTimeMachineBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
            - Success Story
            - Inspiring Story
            - Recognition
            - Donor Support
            - Event by Kids
            - Visitor Visit
            - Challenge / Problem
            - General Update
        home:
          type: string
          enum:
            - All Homes
            - Girls Home - Uppal
            - Blind Home - Begumpet
            - Old Age Home - Peerzadiguda
        eventDate:
          type: string
          format: date-time