# lib/api-zod/src/generated/types/createSponsorshipBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NGO Donor Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSponsorshipBodyStatus } from "./createSponsorshipBodyStatus";

export interface CreateSponsorshipBody {
  donorId: string;
  beneficiaryId: string;
  startDate: Date;
  /** @nullable */
  endDate?: Date | null;
  monthlyAmount: number;
  currency: string;
  status: CreateSponsorshipBodyStatus;
  /** @nullable */
  notes?: string | null;
}