# lib/api-zod/src/generated/types/sponsorship.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NGO Donor Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Sponsorship {
  id: string;
  donorId: string;
  beneficiaryId: string;
  startDate: Date;
  /** @nullable */
  endDate?: Date | null;
  monthlyAmount: number;
  currency: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}